Push

+++ database/migrations/2023_08_13_212736_create_produtos_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->text('codbarra')->nullable();
             $table->string('descricao');
             $table->tinyInteger('movimento');
-            $table->double('valcompra');
-            $table->double('valvenda');
+            $table->decimal('valcompra', 10,2);
+            $table->decimal('valvenda', 10,2);
             $table->tinyInteger('unidade');
             $table->string('estmaximo');
             $table->string('estminimo');
